+++ pages/api/lexio-divinas/index.ts
@@ -1,7 +1,7 @@
 import { CollectionName } from 'constants/mongodb';
 import { InsertOneResult, ObjectId } from 'mongodb';
 import { NextApiRequest, NextApiResponse } from 'next';
-import { LexioDivina, User } from 'types/interfaces';
+import { LexioDivina, User } from 'types/document';
 import { ApiErrorData, isLoggedIn } from 'utils/auth';
 import Mongodb from 'utils/mongodb';
