+++ pages/api/login/kakao.ts
@@ -2,7 +2,7 @@ import axios, { AxiosResponse } from 'axios';
 import { CollectionName } from 'constants/mongodb';
 import jwt from 'jsonwebtoken';
 import { NextApiRequest, NextApiResponse } from 'next';
-import { User } from 'types/interfaces';
+import { User } from 'types/document';
 import { ApiErrorData } from 'utils/auth';
 import Mongodb from 'utils/mongodb';
