+++ types/document.ts
